Split Smart Search palette rendering
Move prompt/result HTML rendering and escaping into a feature-local paletteRender module so palette.ts can focus on wiring events and search behavior.
